+++ JotReverbPatch.hpp
@@ -99,9 +99,9 @@ OWL PATCH:
 #define MAX_CUTOFF				0.4975
 #define MIN_CUTOFF				0.1134
 
-#define SQRT8					2 // sqrt(6)                // 2.82842712474619	// sqrt(8)
-#define ONE_OVER_SQRT8			0.5 // 1/sqrt(6)              // 0.353553390593274	//  1/sqrt(8)
-#define ALPHA					0.873580464736299 // pow(3/2, -1/(6-1))    // 0.943722057435498	//  pow(3/2, -1/(8-1))
+#define SQRT8					2 // sqrtf(6)                // 2.82842712474619	// sqrtf(8)
+#define ONE_OVER_SQRT8			0.5 // 1/sqrtf(6)              // 0.353553390593274	//  1/sqrtf(8)
+#define ALPHA					0.873580464736299 // powf(3/2, -1/(6-1))    // 0.943722057435498	//  powf(3/2, -1/(8-1))
 //       of the 8 delay lines, the longest is 3/2 times longer than the shortest.
 //       the longest delay is coupled to the room size.
 //       the delay lines then decrease exponentially in length.
@@ -161,7 +161,7 @@ typedef struct {
 
 void BuildPrimeTable(char* prime_number_table)
 {
-	int max_stride = (int)sqrt((float)PRIME_NUMBER_TABLE_SIZE);
+	int max_stride = (int)sqrtf((float)PRIME_NUMBER_TABLE_SIZE);
 	
 	for(int i=0; i<PRIME_NUMBER_TABLE_SIZE; i++)
 	{
@@ -300,11 +300,11 @@ void reverbSetParam(reverbBlock* this_reverb, float fSampleRate, float fWet, flo
 		fPreDelaySamples = 0.0;
 	
 	
-	float	fCutoffCoef		= exp(-6.28318530717959*fCutOff);
+	float	fCutoffCoef		= expf(-6.28318530717959*fCutOff);
 	
 	this_reverb->dry_coef	= 1.0 - wetCoef;
 	
-	wetCoef	*= 1.6 * SQRT8 * (1 - exp(-13.8155105579643*fRoomSizeSamples/fReverbTimeSamples));			// additional attenuation for small room and long reverb time  <--  exp(-13.8155105579643) = 10^(-60dB/10dB)
+	wetCoef	*= 1.6 * SQRT8 * (1 - expf(-13.8155105579643*fRoomSizeSamples/fReverbTimeSamples));			// additional attenuation for small room and long reverb time  <--  expf(-13.8155105579643) = 10^(-60dB/10dB)
     //  toss in whatever fudge factor you need here to make the reverb louder
 	this_reverb->wet_coef0	= wetCoef;
 	this_reverb->wet_coef1	= -fCutoffCoef*wetCoef;
@@ -312,7 +312,7 @@ void reverbSetParam(reverbBlock* this_reverb, float fSampleRate, float fWet, flo
 	fCutoffCoef /=  (float)FindNearestPrime(this_reverb->primeNumberTable, (int)fRoomSizeSamples);
 	
 	float	fDelaySamples	= fRoomSizeSamples;
-	float	beta			= -6.90775527898214/fReverbTimeSamples;			// 6.90775527898214 = log(10^(60dB/20dB))  <-- fReverbTime is RT60
+	float	beta			= -6.90775527898214/fReverbTimeSamples;			// 6.90775527898214 = logf(10^(60dB/20dB))  <-- fReverbTime is RT60
 	float	f_prime_value;
 	int		prime_value;
 	
@@ -323,28 +323,28 @@ void reverbSetParam(reverbBlock* this_reverb, float fSampleRate, float fWet, flo
 	this_reverb->delay0.delay_samples	= prime_value - CHUNK_SIZE;									// we subtract 1 CHUNK of delay, because this signal feeds back, causing an extra CHUNK delay
 	f_prime_value			= (float)prime_value;
 	this_reverb->LPF0.a1	= f_prime_value*fCutoffCoef - 1.0;
-	this_reverb->LPF0.b0	= ONE_OVER_SQRT8*exp(beta*f_prime_value)*(this_reverb->LPF0.a1);
+	this_reverb->LPF0.b0	= ONE_OVER_SQRT8*expf(beta*f_prime_value)*(this_reverb->LPF0.a1);
 	fDelaySamples *= ALPHA;
 	
 	prime_value				= FindNearestPrime(this_reverb->primeNumberTable, (int)fDelaySamples);
 	this_reverb->delay1.delay_samples	= prime_value - CHUNK_SIZE;
 	f_prime_value			= (float)prime_value;
 	this_reverb->LPF1.a1	= f_prime_value*fCutoffCoef - 1.0;
-	this_reverb->LPF1.b0	= ONE_OVER_SQRT8*exp(beta*f_prime_value)*(this_reverb->LPF1.a1);
+	this_reverb->LPF1.b0	= ONE_OVER_SQRT8*expf(beta*f_prime_value)*(this_reverb->LPF1.a1);
 	fDelaySamples *= ALPHA;
 	
 	prime_value				= FindNearestPrime(this_reverb->primeNumberTable, (int)fDelaySamples);
 	this_reverb->delay2.delay_samples	= prime_value - CHUNK_SIZE;
 	f_prime_value			= (float)prime_value;
 	this_reverb->LPF2.a1	= f_prime_value*fCutoffCoef - 1.0;
-	this_reverb->LPF2.b0	= ONE_OVER_SQRT8*exp(beta*f_prime_value)*(this_reverb->LPF2.a1);
+	this_reverb->LPF2.b0	= ONE_OVER_SQRT8*expf(beta*f_prime_value)*(this_reverb->LPF2.a1);
 	fDelaySamples *= ALPHA;
 	
 	prime_value				= FindNearestPrime(this_reverb->primeNumberTable, (int)fDelaySamples);
 	this_reverb->delay3.delay_samples	= prime_value - CHUNK_SIZE;
 	f_prime_value			= (float)prime_value;
 	this_reverb->LPF3.a1	= f_prime_value*fCutoffCoef - 1.0;
-	this_reverb->LPF3.b0	= ONE_OVER_SQRT8*exp(beta*f_prime_value)*(this_reverb->LPF3.a1);
+	this_reverb->LPF3.b0	= ONE_OVER_SQRT8*expf(beta*f_prime_value)*(this_reverb->LPF3.a1);
 	fDelaySamples *= ALPHA;
 
 }
@@ -383,8 +383,8 @@ void Delay(delayBlock* this_delay, float* input)
 //
 //	transfer function:			H(z) = b0/(1 - (a1+1)*z^(-1))
 //
-//	for the nth delay line:		a1   = delay[n]/size * exp(-2*pi*fcutoff/Fs) - 1  =   pole - 1
-//								b0   = a1/sqrt(8) * 10^(-(60dB*delay[n]/RT60)/20dB)
+//	for the nth delay line:		a1   = delay[n]/size * expf(-2*pi*fcutoff/Fs) - 1  =   pole - 1
+//								b0   = a1/sqrtf(8) * 10^(-(60dB*delay[n]/RT60)/20dB)
 //
 void Filter(filterBlock* this_filter, float* input)
 {
